pyproject.toml
[project]
name = "ask-my-file"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.1.10",
    "langchain-community>=0.0.25,<0.1",
    "langchain-groq==0.0.1",
    "langgraph==0.0.26",
    "streamlit==1.31.0",
    "python-dotenv==1.0.0",
    "docx2txt==0.8",
    "duckduckgo-search==4.4.3",
    "faiss-cpu==1.7.4",
    "sentence-transformers==2.3.1",
    "pymupdf==1.23.21",
    "python-pptx==0.6.23",
    "pandas==2.2.0",
    "openpyxl==3.1.2",
    "tiktoken==0.6.0",
    "numpy==1.26.4",
    "pydantic<2.0",
    "doc2text>=0.2.4",
]
